relocate PostDetailFragment
