+++ app/src/main/java/csis3175/w23/g11/rooftown/posts/ui/view/ListViewFragment.java
@@ -19,7 +19,6 @@ import com.google.android.material.bottomnavigation.BottomNavigationView;
 import java.util.ArrayList;
 import java.util.UUID;
 
-import csis3175.w23.g11.rooftown.PostDetailFragment;
 import csis3175.w23.g11.rooftown.R;
 import csis3175.w23.g11.rooftown.posts.ui.adapter.ListAdapter;
 import csis3175.w23.g11.rooftown.posts.ui.viewmodel.PostViewModel;
