+++ app/src/main/res/layout/fragment_posting.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PostDetailFragment">
+    tools:context=".posts.ui.view.PostDetailFragment">
 
     <ImageView
         android:id="@+id/imgViewPosting"
